# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deep Control Loads Backend Rest API Documentation",
        "description": "Rest API Documentation",
        "contact": {
            "email": "remstroyod@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/orders/{order}/documents/{document}": {
            "delete": {
                "tags": [
                    "Admin: Orders"
                ],
                "summary": "Delete Document Rest API",
                "description": "Delete Document",
                "operationId": "admin.orders.documents.destroy",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID Document",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Get All Users Rest API",
                "description": "Get All Users",
                "operationId": "admin.users.index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Store User Rest API",
                "description": "Store User",
                "operationId": "admin.users.store",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User Name",
                                        "type": "string",
                                        "example": "Manager Name"
                                    },
                                    "surname": {
                                        "description": "User SurName",
                                        "type": "string",
                                        "example": "Manager Surname"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "manager@example.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "gender": {
                                        "description": "User Gender",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "description": "User Status",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "description": "User Role",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "users": {
                                        "description": "Users Attach Manager",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}": {
            "put": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Update User Rest API",
                "description": "Update User",
                "operationId": "admin.users.update",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "User Status",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "description": "User Role",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "users": {
                                        "description": "Users Attach Manager",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Delete User Rest API",
                "description": "Delete User",
                "operationId": "admin.users.destroy",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/managers/all": {
            "get": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Get All Managers Rest API",
                "description": "Get All Managers",
                "operationId": "admin.users.managers.index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/managers/reattach/{userFrom}/{userTo}": {
            "post": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Reattach Users Manager to Manager Rest API",
                "description": "Reattach User Manager to Manager",
                "operationId": "admin.users.managers.reattach",
                "parameters": [
                    {
                        "name": "userFrom",
                        "in": "path",
                        "description": "ID Manager From",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "userTo",
                        "in": "path",
                        "description": "ID Manager To",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/managers/attach/{manager}/{user}": {
            "post": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Attach User to Manager Rest API",
                "description": "Attach User to Manager",
                "operationId": "admin.users.managers.attach",
                "parameters": [
                    {
                        "name": "manager",
                        "in": "path",
                        "description": "ID Manager",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/{user}/orders": {
            "get": {
                "tags": [
                    "Admin: Users"
                ],
                "summary": "Get All Orders by Users Rest API",
                "description": "Get All Orders by Users",
                "operationId": "admin.users.orders",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page the number of returned orders",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/orders/{order}/documents/{document}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete Document Rest API",
                "description": "Delete Document",
                "operationId": "orders.documents.destroy",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID Order",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID Document",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}